+++ lib/constants/constants.dart
@@ -1,7 +1,7 @@
 /*
  * BSD 3-Clause License
 
-    Copyright (c) 2020, RAY OKAAH - ray@flutterengineer.com
+    Copyright (c) 2020, RAY OKAAH - MailTo: ray@flutterengineer.com, Twitter: Rayscode
     All rights reserved.
 
     Redistribution and use in source and binary forms, with or without
@@ -38,3 +38,4 @@ const URL_JWT_TOKEN = '$URL_JWT_BASE/token';
 const DEFAULT_WC_API_PATH = "/wp-json/wc/v3/";
 const URL_WP_BASE = '/wp-json/wp/v2';
 const URL_USER_ME = '$URL_WP_BASE/users/me';
+const URL_REGISTER_ENDPOINT = '$URL_WP_BASE/users/register';